Updated to my current test build
Updated to my current test build.

Added a test folder for scripts I'm playing around with.

Changed how the debug and markers work.
You will need to open the .sqf's and add your playeruid's.

+++ HC/Debug/zMarkers.sqf
@@ -1,39 +1,44 @@
-if(isNil "markers") then { markers = []};
+_uid = ["0"];
 
-ZombieVisibleDistance=15000;
-ZombieMarkerType="vehicle";
-ZombieMarkerColor="ColorRed";
-ZombieName="Z";
+if (getPlayerUID player in _uid ) then { //Debug
+
+    if(isNil "markers") then { markers = []};
+
+    ZombieVisibleDistance=15000;
+    ZombieMarkerType="vehicle";
+    ZombieMarkerColor="ColorRed";
+    ZombieName="Z";
 
 
-while {true} do {
-    _pos = getPos player;
-    _zombies = _pos nearEntities ["zZombie_Base",ZombieVisibleDistance];
-    _zmcount= count _zombies;
-    ZM=0;
+      while {true} do {
+        _pos = getPos player;
+        _zombies = _pos nearEntities ["zZombie_Base",ZombieVisibleDistance];
+        _zmcount= count _zombies;
+        ZM=0;
 
-    _markcount = count markers;
-    for "ZM" from 0 to (_markcount -1) do
-    {
-        deleteMarkerLocal ("zmMarker"+ (str ZM));
-    };
-
-    for "ZM" from 0 to _zmcount do {
-
-        _text = format ["zmMarker%1", ZM];
-        markers set [ZM, _text];
-        zmB = _zombies select ZM;
-        if(alive zmB) then {
-            posTT = position zmB;
+        _markcount = count markers;
+        for "ZM" from 0 to (_markcount -1) do
+        {
             deleteMarkerLocal ("zmMarker"+ (str ZM));
-            MarkerZm = "zmMarker" + (str ZM);
-            ParamsZm=[MarkerZm,posTT];
-            MarkerZm = createMarkerLocal ParamsZm;
-            MarkerZm setMarkerTypeLocal ZombieMarkerType;
-            MarkerZm setMarkerPosLocal (posTT);
-            MarkerZm setMarkerColorLocal(ZombieMarkerColor);
-            MarkerZm setMarkerTextLocal ZombieName;
         };
+
+        for "ZM" from 0 to _zmcount do {
+
+            _text = format ["zmMarker%1", ZM];
+            markers set [ZM, _text];
+            zmB = _zombies select ZM;
+            if(alive zmB) then {
+                posTT = position zmB;
+                deleteMarkerLocal ("zmMarker"+ (str ZM));
+                MarkerZm = "zmMarker" + (str ZM);
+                ParamsZm=[MarkerZm,posTT];
+                MarkerZm = createMarkerLocal ParamsZm;
+                MarkerZm setMarkerTypeLocal ZombieMarkerType;
+                MarkerZm setMarkerPosLocal (posTT);
+                MarkerZm setMarkerColorLocal(ZombieMarkerColor);
+                MarkerZm setMarkerTextLocal ZombieName;
+            };
+        };
+        sleep 3;
     };
-    sleep 3;
 };
